fix probability() in Decile.kt Try using optimalComparison for Boulder CLCA. Remove RunTestRepeatedResult.percentHist Remove SRT.percentHist Move Deciles to TestDeciles. Rename Quantil to Deciles in util.
